chore: refactor from PR comments
