+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/ProgramStageSelector/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { ProgramStageSelector } from './ProgramStageSelector.container';
