+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/TopBar/topBar.types.js
@@ -6,8 +6,8 @@ export type Props = {|
     enrollmentId: string,
     teiDisplayName?: string,
     trackedEntityName?: string,
-    stageName: ?string,
-    eventDateLabel: ?string,
+    stageName?: string,
+    eventDateLabel?: string,
     enrollmentsAsOptions?: Array<Object>,
     onSetOrgUnitId: (orgUnitId: string) => void,
     onResetOrgUnitId: () => void,
