--- src/core_modules/capture-core/components/WidgetProgramStageSelector/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { WidgetProgramStageSelector } from './WidgetProgramStageSelector.container';
